+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - [Provide an Oauth2 scope for the staff role #127](https://github.com/farmier/farm_rcd/issues/127)
+- [Enable staff to comment on plans #126](https://github.com/farmier/farm_rcd/issues/126)
 
 ## [1.2.0] 2026-01-21